Added the vectoradd case (C++ bindings test). It crashes.
The pthread and native devices are now both CPUs and GPUs to enable
more test cases out of the box. The reasoning is that with the multi-WI
vectorization the CPUs should not be treated as "task parallel intra-WI"
targets like some people seem to assume of CPUs vs. GPUs but more
like data parallel devices such as GPUs are commonly.