Another attempt for fixing #938123. The problem wasn't in replicating the code but replicating the variables. The inner loop referred to an outside-loop gid which wasn't corrected to the replicated value because that region wasn't handled yet.
Fixed it to first produce the ParallelRegions so it finds all
variables after which to start replication.

Another issue was with the loop with volatile iteration variable which
created a construct that allowed the loop be skipped entirely (i = 0;
check; loop;). There was a barrier missing from the loop skip check
which caused multiple replication to happen which exploded the thing.

There's one XFAIL still in ./testsuite -k regression