Fixes the case of tail replication before WI ID dependent code. It's not correct to always add a barrier to the split block because the condition can be ID dependent.
Reverted back to the old behavior of conditional barrier replication and
fixed a bug that appeared with that: when a BB is duplicated multiple
times, like happens with conditional barriers the variables defined
in the BB are updated to the reference map only once, thus overwriting
the old clone's variables. This causes some users to use the wrong
variables (not dominated). I fixed this by appending a prefix to the
multiply cloned BBs and fix the undominated references by trying the
alternative. Currently the kludgish (but quite fast) way is to just
try with and without the added .pocl_1 suffix, in case the instruction
has an operand that doesn't dominate it.