Always inline all non-kernel functions in Flatten, regardless if they use the magic variables or not.
Drop uncalled functions (including kernels we are not generating a WG for) 
by calling globaldce. This results also in not processing the kernels we
are not interested by the latter passes in the WG generation. Should speed up 
kernel compilation and code generation, reduce "noise" when debugging and
reduce the final binary size.