The case (marked as XFAIL) 21: for-loop with computation after the brexit now replicates a kernel crash occuring in ViennaCL's examples/tutorial/fft.
An issue revealed by this case was that it traversed a back edge in this
case, replicating too much in the BTR due to faulty loop backedge
detection. Replaced it with a dominance-based check that fixed it. Anyways,
the actual crash is not directly caused by this but it's caused by the
fact that the replicated tail is the exit part of the inner loop, jumping
to the outer loop's header, thus creating a new edge to the loop header.
This requires the phi nodes to be fixed to account for the new variable
alternatives.