Allow to run tests when pocl is compiled in ICD mode only
In ICD mode only, programs must be run through the (already) installed ICD loader
(ie through the system libOpenCL.so)
Of course, this system library uses /etc/OpenCL/vendors to find available ICD, but:
- currently built POCL is not there
- putting something there require root access
- programs need to choose the correct (POCL) platform if we want to test it

ocl-icd allows to define the OCL_ICD_VENDORS environment variable
in order to change the default /etc/OpenCL/vendors directory.
So, the configure script create a specific ICD file in
$(builddir)/ocl-vendors/pocl-tests.icd
and atlocal is used to define OCL_ICD_VENDORS for tests.

So, tests can be run even when POCL is configurated in ICD mode only
at the condition that the ICD loader is ocl-icd.

Note that ocl-icd version 1.4 at least is required
(or the Debian package 1.3-3). Previous versions were buggy.