Added code to split large BBs so the dot doesn't choke on the CFG dump. This helped to track the miscompilation reason. It's caused by an illegally handled PHINode in the beginning of the region.
In prod_AA of ViennaCL, Csub is created from a PHINode in the beginning of 
the region *but* it's not a loop variable/work group variable but the most 
important work item variable of the kernel. It needs to be replicated.
Somehow.

Do not context save scalar loads. They should not be work item
dependent. Unless the above issue is not fixed by replicating the 
scalarized PHI somehow like it was done at some point. Need to figure 
it out.