Use fixed address space id for all but local address space in _kernel.h. This was reverted for some reason at some point.
This is to ensure we do not lose the information of the
OpenCL address space in the Clang for targets (basically regular CPUs)
that map all them to 0. 

local is left as is because it needs special handling in Clang.
Automated local variables must be converted by it to special global
variables which are handled by pocl kernel compiler passes.

This mess will be cleaned up at latest when SPIR and its standard
address space numbers gets finished and implemented in Clang. 

This caused the issue with bfs / Parboil of not finding the atomic function 
for AS3 pointer. Now it runs successfully and is enabled in the suite.