benchmark.py can now exclude POCL kernel compiler time by retaining the compilation results between repeats which makes the latter executions to reuse the old ones.
This seems to help some of the cases which means that the AMD cases do not 
actually use the profiling queue afterall (it seems) but a gettimeofday timer 
which includes the compilation times too.

Also set POCL_BUILDING=1 when running the benchmark.py to test the pocl from the
build tree.