Reverted the OpenCL header update for now as it opened a can of worms. You can merge it back by 'bzr merge -r618..617 lp:pocl'.
The C++ API implemented regression tests started to crash. The property
loading failed and it tried to free an illegal property struct in 
clGetContextWithType(). However, I'm not sure if it was because of mixed
versions (icd loader 1.1) or some actual issue revealed by the
new headers.

The ICD loader has to be updated to 1.2 so we need to warn developers 
before committing this change. E.g., clReleaseDevice() is not in 1.1
so all cl.hpp cases fail because they have a reference to it and it's
not in libOpenCL.so.

Another thing I noticed is that the regression tests link against pocl
twice and they end up depending on an installed pocl due to libpoclu
having pocl as a dependency. It should not link against it in icd mode
but to -lOpenCL. I'll fix this separately.