Fixing the regression for the fast launcher revealed a problem with the byval handling in general. The byval args point to the default address space and we assume them to be in the global. For flat address space targets (read CPUs) this makes no difference but heterogeneous setups fail.
Now that the fast work group launcher tries to call the kernel assuming
all args are in the global space, it crashes as it does not find a
correct finger print.

This should be fixed properly when we adopt SPIR as the input. Then the
all the args should be in pointers (to global, hopefully). 

Marked the vectorarg test as an XFAIL for now, until there's more time
to look into it.