Ensure the local id loads are always done inside the parallel region.
This fixes a bug in a couple of AMD cases introduced after optimizing
the kernel library (get_local_id etc.). In some cases when a local id
was stored to a temporary variable in a separate parallel region,
a wrong local id was used when executing another region with the
wiloops method.