Fixed a kernel compiler bug when handling a b-loop with multiple latches.
The problem was threefold:

1) Detection of the need to "peel" the wiloop was faulty: post-dominance
check does not work in this case as the another branch leads always to
exit via the back edge, thus is post-dominated even though it should be
handled as a diverging region. Fixed this by counting the number of
parallel regions sharing the entry block and assuming if there are more
than one, the PR is a diverging one and should be peeled.

2) The parallel region was not found for the another latch created by
the barrier tail replication. It assumed there can be only one latch
barrier.

3) Barrier tail replication failed to replicate the latch due thinking
the PHI node values should all come from the replicated block. The
PHI node values in this case originate from the "shared" part of
the replicated diverging parallel region so there was no own copy for
the variable in the replicated part. Now it simply falls back to using
the original one in this case.