Make "make dist" work again
Include/exclude many files that should or should not be included into
the tarball generated by "make dist".

Vecmathlib is expected to be present, and is included in the tarball.

The kernel library sources (those listed in sources.mk) are handled
via shell glob patterns. This is not elegant, but I couldn't find a
good way to get these variables out of sources.mk.