configure: check for aligned allocator in C mode
The components that use pocl_aligned_malloc are all C files, so check
for availability of aligned_alloc() or posix_memalign() using the
(properly set-up) C compiler, not the C++ compiler.

This prevents compiler errors and warning raised when one of the
functions is available when compiling with C++ but not when compiling in
C (without additional flags), which is e.g. the case in the current
Debian unstable.