Clean up get_supported_image_formats
When gathering information from devices, request a const ptr, since this
is what we get in return in meaningful cases (basic device). This is
more sensible anyway (we don't manipulate the devices) and removes a
compile-time warning.

While we're at it, also clarify the behavir during information gather in
the case when one device returns no supported image formats.