Correct isequal, isnan, and similar routines
isequal, isnan etc. need to return +1 for "true" for scalar arguments,
but -1 for "true" for vector arguments. Do this.

Introduce new templates with type signatures L_VV and L_V, where L
denotes an integer expressing a truth value (with +1 or -1 for true as
described above).

Simplify implementation and improve performance for two-argument
functions.