Revamp handling target triples and CPUs
It is necessary to specify both target triple and CPU to define the ABI. If not, e.g. x86-64 (AVX support) or ARMv6 (hardware floating-point support) break, leading to inconsistencies between the kernel library and compute kernels.

In the new approach, configure.ac defines command line options CLANG_FLAGS and LLC_FLAGS that contain all necessary flags to specify target triple and CPU. It is not necessary any more to explicitly specify a target (in fact, these are now ignored).