Add POCL_KERNEL_CACHE_IGNORE_INCLUDES env
By default, the kernel compiler cache does not cache kernels that
have #include clauses. Setting this to 1 changes this so that the
includes are ignored and not scanned for changes. Use this to
improve the kernel compiler hit ratio in case you know that the
included files are not modified across runs.