Clip max_work_item_sizes to max_work_group_size
CL_DEVICE_MAX_WORK_ITEM_SIZES is an array describing the maximum
number of work-items that can be indexed in a given dimension in a
work-group (not globally), so it doesn't make any sense for it to be
larger than CL_DEVICE_MAX_WORK_GROUP_SIZE, which is the overall limit
in the number of work-items in a work-group.

Just as well, since we don't have hardware work-item indices (which is
the case on GPUs instead) it also doesn't make any sense for any of the
values to be _less_ than the maximum work-group size (i.e. there's no
reason to not allow 1×1×MAX as well as MAX×1×1 and 1×MAX×1), so just
set each element of max_work_item_sizes to the value of
max_work_group_size.

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>