Always clone event list when creating command
This avoids ending up with cycles in the dependency graphs in case the
caller decides to recycle the same array before we're done using it.

The change is less significant that it might seem because we did a clone
anyway when adding the previous command if it wasn't done yet.