Make the kernel cache depend on preprocessed CL source, part 1
This is a series of patches to let pocl cache on SHA1 of the
preprocessed program source, instead of just program source,
which is prone to mess up caching via macros / #include etc.

This involves a rather big change in pocl caching, since it
turns out program.bc depends on device as well, so these
patches change per-program cachedirs into per-device+program
cachedirs.

Changes in this patch:
* struct _cl_program: remove cache_dir and cachedir_lock, they
  don't make sense with per-device cache dirs
* struct _cl_program: add build_hash array (for per-device SHA1)
  and build_log array (pocl_llvm_build_program might run into
  errors before it gets the preprocessed source)
* change clCreateProgram* and clReleaseProgram  accordingly