Correct handling of subdevices in clBuildProgram()
Currently the subdevices implementation is a "stub" and
the subdevice's cl_device_id equals to its parent's. So we
count the number of real devices (for each subdevice, only
need to build the program once, for its parent device) and
error only if we built less real devices than requested.