Command queue management utility
To prevent segmentation faults and other crashes when applications end,
we have to ensure that all command queue are finished before any of the
automatic destructors trigger. Otherwise, it can happen that e.g. some
LLVM globals are freed before the command queue are released, and
the subsequent release of the command queue can cause a kernel
compilation and then crash or hang inside LLVM. This is particularly
noticeable in some configuration with the ViennaCL examples.

We solve this by keeping track of all the allocated command queues and
forcing a clFinish on them at exit (which is called before the
automatic destructors).