Autonomous POCL_UPDATE_EVENT_* macros
The macro don't need an explicit command queue passed any more, just a
pointer to an event. The command queue is taken to be the event's own
command queue, if present. The existence of the command queue is
checked before dereferencing.

(I'm not 100% sure that the check is necesary, since I _think_ the only
events that do not have an associated command queue are user events,
which shouldn't be touched by these macros, so an assert might be
sufficient.)