Add context member to _cl_event
Automatic events all have an associated command queue, and this was
used to retrieve the context in clGetEventInfo. However, user events
have a NULL command queue, so this approach would lead to segmentation
faults.

Since user events still have an associated context, add the actual
context to the event structure.