Make program cachedir creation more robust
We must be able to create a cachedir even in the case of preprocessing
failure, so handle this case by hashing the original source code if no
preprocessing text is given.

Also, binaries should be invalidated whenever the hash changes, so move
the binary invalidation inside the cachedir creation function.