[cmake] Make it possible to specify the host CPU.
Mips currently autodetects LLC_HOST_CPU="(unknown)" which is not a valid CPU
name. This had the side effect of disabling cl_khr_int64, and removing
floating point support from printf().

This patch removes the unconventional caching mechanism for LLC_HOST_CPU in
favour of cmake's usual method which allows the user to specify it with
-DLLC_HOST_CPU=...