[cmake] Make it possible to override CL_DISABLE_HALF.
I suspect the check is currently wrong. It seems to be checking whether C
allows half precision types in arguments and returns. On Mips, this is not
currently allowed. However, I suspect OpenCL allows this regardless of
target.

This patch removes the unconventional caching mechanism for
CL_DISABLE_HALF in favour of cmake's usual method which allows the user
to specify it with -DCL_DISABLE_HALF=ON/OFF. This allows me to investigate
whether the check is correct or not.