Revert "Don't cast to or from (void *), especially not malloc returns"
This reverts commits 4b458c77ac6976322378cd85f96e8ff52c6b1f87 and
4c61d8734c607d98bf8c3404a2094193f4d5cbf3

The reason to revert is that, on Windows, pocl is compiled with MSVC
which only has a C++ compiler, and it's an error to NOT cast the
result of malloc in C++:
https://stackoverflow.com/questions/16815476/why-must-i-cast-the-result-from-malloc

Also, any modern GCC will emit an "implicit declaration" warning when
you use malloc() without including <stdlib.h> regardless of having or
not having a cast. So the only problem that remains is the clutter.