pocl-cache: refactor pocl_cache_work_group_function_so_path
Currently pocl_llvm_generate_workgroup_function is checking
for parallel.so in cache, instead of parallel.bc. HSA does
not ever produce .so files, so this check fails and triggers
pocl_llvm_generate_workgroup_function is re-run everytime,
when using HSA.

This patch fixes the issue by changing the check to parallel.bc.