Refactor clEnqueueCopyBufferRect and clEnqueueCopyImage
clEnqueueCopyImage used to call clEnqueueCopyBufferRect, which couldn't
work because the latter would check that the source and destination
were actual images.

Refactor the code into a separate function, with extra parameters to
know which should be a image and which a buffer, and copy it from both
the API calls instead.