HSA: 2nd version of OoO driver for HSA devices, fully asynchronous
Uses a single separate pthread for actual work. Driver pthread uses
hsa_amd_signal_wait_any() or a fallback hsa_signal_wait() to wait.
Many things are not optimal, a few unimplemented, but most of the
AMD SDK samples that used to work with the old driver now pass too.