TAS: handle OpenCL 2.0 generic address space with LLVM 3.8
Note that "handle" above does not mean it's working as intended in
OpenCL 2.0 standard, this merely converts "hidden" generic pointers
(by OpenCL 2.0 standard all pointer without specified AS point to
generic AS) to pointers to private AS (the default behaviour before
OpenCL 2.0). This is enough to compile 1.2 programs with -cl-std=2.0
compile flag, but obviously does not properly handle generic AS.

Proper solution will require much more work on LLVM and pocl.