TAS: fix generic AS on prototype arguments
generic AS was not removed from prototypes. This was mostly not
a problem since prototypes get resolved when linked into parallel.bc,
but some (standard functions e.g. fflush() and vprintf(), mainly
called from our OpenCL printf implementation) are not resolved,
and this was causing issues with LLVM 3.8, when the actual call
has a non-generic-AS argument but the prototype still has generic AS
in argument list. LLVM would fail with error like this:

ValueEnumerator.h:132: unsigned int
llvm::ValueEnumerator::getTypeID(llvm::Type*) const:
Assertion `I != TypeMap.end() && "Type not in ValueEnumerator!"' failed.