Add undocumented POCL_FORCE_PARALLEL_OUTER_LOOP env.
This env can be used to temporarily control the logic whether to attempt
injecting the parallell WI loop inside the kernel loops which has the
effect of enabling 'outer loop vectorization' across the work-items
inside the user's loops.

The default behavior is that this is not attempted in case the kernel
doesn't contain any barriers because adding the explicit barriers that
cause the outer loop vectorization also can add overheads in form of
work-item context save/restore code.