TAS: do not reconstruct kernel metadata for non-SPMD devices
The opencl.kernel metadata needs to be maintained only for
SPMD devices (read: HSA at the time of this writing) because
for non-SPMD machines the original kernel function is not
there anymore, but is replaced with a work-group function.

The reconstruction code needs to be ported to LLVM 3.8+, but
hopefully at that point we have gotten rid of the TAS mess altogether.