Support for building without the fake address spaces (#394)
Relies on the included patch to clang that makes it produce SPIR address
spaces in the kernel arg info metadata. It's pending review
in https://reviews.llvm.org/D26157  It applies at least to LLVM 3.9 and
LLVM trunk. Possibly also to earlier versions.

Cmake detects if the patch is applied and in that case disables the
use of fake address spaces everywhere and detects the kernel argument
address space from the metadata instead.

This will be the default in the future as soon as the patch gets applied
to LLVM upstream.

Also snuck in some misc. cleanups while going though the code.