Avoid the bitcode linker exploding the bitcode size
Two issues caused this (which were visible only with Clang 4.0):
- bitcast() template of pocl-compat was static which caused a copy
of the function created for each bitcode file which was finally
linked to the big one, and the copies followed. Now it's global
which should allow merging the definitions.
- There was a bug in the linker which caused it to copy functions
that were referred from multiple places multiple times by
appending the new definition to the end of the old definition.
This exploded the size of some of the helper functions
copied over.

Fixes #433.