Fix printf for Clang 4.0.
The problem was that Clang does not allow vararg functions other that
printf in OpenCL C mode (mandated by the specs).

Fixed this by introducing only the printf declaration in .cl mode and
converting the calls at bitcode linker to the internal _cl_printf().

Also had to clean up the build a bit as it enabled OpenCL C mode even
for C builtin library files due to the OpenCL C standard switch being
passed to it.

Fixes #440.