Workaround #439 (EinsteinToolkit hangs with LLVM 4.0)
The hang happens at the WG function inlining to the launcher. The
inliner's AddAliasScopeMetadata runtime explodes, likely due to
the large number of restrict kernel args in combination with a
huge number of lifetime markers. Remove the markers before
inlining to speed it up until a proper fix lands.

Fixes #439.