Correctly create command queue
As currently implemented the command queue creation with properties always
fails (so long as there is an item set in the CL_QUEUE_PROPERTIES). I think the
original intention of the code was something like what I'm proposing here, so
here it goes. I've tested this change locally and now I get a queue created
indeed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>