clEnqueueNDRangeKernel: new NULL local size logic
When the user does not specify a local work size, the platform should
decide one on its own.

The local size is subject to the following _hard_ constraints:

1. in each dimension, the corresponding local size must divide the
corresponding global size;
2. in each dimension, the corresponding local size must not be larger
than the corresponding device limit;
3. the overall local size (product of the local sizes in each
dimension) must not be larger than the maximum overall local size
allowed by the device.

Additionally, we would like the local size to be a multiple of the
device-recommended work-group size multiple for the kernel, to improve
device usage or vectorization opportunities. Note that this is not a
hard constraint, but rather a performance hint.

The logic previously used to determine the local size did not take into
account the hard constraints 1 and 2: it started from the global size as
potential local size and then reduced the three dimensions (by halving
or clipping) until the local size satisfies constraint 3, while trying
to satisfy the performance hint too..

This could result in an invalid local size (violating either 1 or 2
or both), which would then be caught in the subsequent checks (that
hve now been moved to the the user specified local size in the previous
commits).

This commits rewrites the logic in order to aim for correctness first,
by taking constraints 1 and 2 into account when looking for an initial
candidate for the local size. A reduction quite similar to the existing
one is then effected on the local size candidate until constraint 3 is
satisfied.

The main differences compared to the old logic are:

* to satisfy constraint 1, the grain size for the performance hint may
be a fraction of the device preferred work-group size multiple;

* in the reduction phase, we always reduce the third dimension first,
then the second dimension and finally the first one; this may result in
the new logic providing local sizes with a different shape (but the
same overall size) than the old logic, in cases where the old logic
would produce valid results.

The new logic is still not necessarily more optimal (performance-wise)
than the old logic, but is guaranteed (modulo bugs introduced by yours
truly) to produce valid local sizes, and should not introduce any
meaninful performance regressions related to overall block sizes (e.g.
concerning CU usage), but there might be shape-related performance
regressions, for kernel/device combinations where more square work-group
sizes would be preferrable.

Performance-related local size optimizations are deferred to future
patches, that may take device-specific and kernel-specific features for
local size tuning (hopefully without breaking correctness).