clEnqueueNDRangeKernel: try to fill all compute units
Our greedy algorithm for NULL local size may cause an imbalanced
workload, with a few compute units bearing the grunt of the work with
oversized work-groups, and the other compute units doing nothing.

We now try to achieve a better balance by further splitting the local
size so that all CUs get at least a work-group, as long as this doesn't
mean making work-groups too small.