[CUDA] Implement global offsets
Given that these aren't frequently used, we only generate a kernel
with support for global offsets if there is an enqueue with non-zero
offsets. The kernel library implementations of get_global_id and
get_global_offset have placeholder global variables for the offsets,
which are replaced with kernel arguments if offset support is
required, or constant zero otherwise.