pocl_topology: actually cleanup behind ourselves
With 351d727ceba5f2797e932c5ad653753e21166de2 we introduced a label for
a cleanup on exit in pocl_topology_detect_device_info, however, this
wasn't actually used (erroneously).

Do jump to it if hwloc_topology_load() fails, as the hwloc documentation
suggests the topology should get destroyed anyway in this case.

Signed-off-by: Giuseppe Bilotta <giuseppe.bilotta@gmail.com>