llvm/Flatten: switch back to force-inlining only some funcs
The new code in Flatten added AlwaysInline attr to everything
except kernel.

This switches Flatten back to its original code (adding AlwaysInline
only to users of global variables); additionally it adds AlwaysInline
on barrier().

The problem with flattening all code is that OpenCL spec requires
maximum number of image arguments to a kernel to be at least 127.
write_imagef() is relatively heavy on stack, and AlwaysInline on a
kernel with 127 images and 127 calls to write_imagef() will require
a giant stack, if we force inlining of all write_imagef() calls.