Linux/x86: add a SIGFPE handler.
The OpenCL spec says:

> A divide by zero with integer types does not cause an exception
> but will result in an unspecified value.

Unfortunately x86 doesn't seem to have any flag to disable SIGFPE
from integer division by zero. To make things worse, the (I)DIV x86
instruction does not advance the IP register when the exception
is raised.

This adds a SIGFPE handler that tries to handle the situation by
decoding the instruction and skipping to the next instruction. In
case the instruction is not recognized as (I)DIV, the original
handler is invoked.