Fix LLVM CPU fallback
This makes it possible to use pocl with LLVM which doesn't
recognize your CPU, as long as you provide a sensible value
for -DLLC_HOST_CPU when you build pocl with CMake.

Also, turn a CMake warning into an error and do not fallback
to "generic" cpu if LLC can't determine the cpu model, because
"generic" cpu doesn't seem to work. Fixes #512