Fix SPIR bitcode for example1-spir32/64
Since linking kernel library now happens in pocl_llvm_build_program
stage, we expect all binaries to be fully linked. This broke SPIR
examples, so this patch updates the examples with fully linked SPIR
binaries.

The real fix would be to link the SPIR binaries at clBuildProgram,
but we need an actual SPIR kernel library first to do that (the
previously used hack of linking to host library is too fragile).