fix issue #537 (pocl_binaries are not functionnal)
We need to skip program.bc in a pocl_binaries when we want to have
information on kernels from the start of a pocl_binaries.
When running POCL offline, we do not need to take care of
program->binaries (even if they are present) as we do not have a
compiler available.