libclc: isnan(), isfinite(), isinf(), isnormal()
These implementations do not depend on clang builtins.

This is not strictly necessary, merely a workaround for a bug
likely in some LLVM 4 optimization pass, where it fails to
compile the conformance tests for isnan() / nan() vector size 8.
Other vector sizes and LLVM 5 work with the code which uses
builtins.