Fix #289: pocl aborts with "Could not create topdir for cache"
As I've described in a comment, #289 is not a pocl issue,
but it might be better to not call abort() in this case,
because the user's program might be able to fallback (using
other OpenCL platform, or using something else entirely).
This commit changes pocl to return CL_DEVICE_NOT_FOUND (it
will still log the error).