pthread driver: use spinlocks on work queue
The wq_lock (work queue lock) in pthread driver was
used for both wake_pool cond variable (signaling when
there was new work to be done) and also to lock access
to kernel/command queues.

The problem with this approach is, with many threads
there would be high lock contention on wq_lock; the
operation would block by doing a syscall, at which
point the linux scheduler would schedule it off the
CPU. This resulted in pthread_scheduler_get_work()
often wasting ~10 ms just to take a look at two
pointers.

This commit adds a separate lock for the wake_pool
cond variable, and another spin lock for the cmd/kernel
queue variables. The spin lock should be a safe
option since it's only ever protecting simple pointer
checks.

Similarly, kernel_run_command->lock was changed to a
spinlock.