Fix #582: undefined symbol: __gnu_f2h_ieee
If the CPU did not have F16C extension, the builtin
vector convert would seem to fallback to some library function
which doesn't exist (at least for CPU target).

This replaces the store/load of halfs in images by vload/vstore
which should handle this properly. Unfortunately it seems
to have some issues with addrspaces on LLVM <= 3.7, so it's only
enabled for 3.8+