Disable use of _Generic keyword for LLVM 6
This is a workaround for a bug in LLVM which doesn't
seem like it'll make it into the 6.0 branch in time.
Pocl issue #599, llvm bug #35826

Unfortunately the fallback expression ((T)0.1f > (T)0.0f)
does not work, and likely never worked (AFAICT in a constant
expression, floats must always be cast to an integer type).