Enables -O1 on compilation of all kernel library files
One of the changes in LLVM 5 changed -O0 (the default)
so that it adds "noinline" keyword to all functions,
which meant subsequent optimization was mostly useless.

Adding -O1 negates the noinline keyword. The problem
was mostly visible when compiling with VML, since
the new SLEEF / libclc was mostly OpenCL code.

There were two issues, one is that with -O1 LLVM
fails to compile lib/kernel/printf.c because
it seems to have a problem with 3-sized vector in
va_args. Changing vector3 load to vector4 load seems
to work around the issue. The other is that CUDA
failed two tests; adding noduplicate to barrier
function fixed it.