Workaround a problem with metadata cloning in linker()
The way metadata are handled in linker, causes some
metadata to be present multiple times. This is *probably*
not a problem, except one instance - "wchar_size" seems
to be copied twice into llvm.module.flags. This caused
verification to fail, caused problems with llvm-as/llvm-dis
and LLVM since 6.0 started to fail with 'example_cherk' from
clBLAS. This workarounds the problem, but most likely is
not a final fix for metadata handling in the linker.