Fix CPU detection
Previously, if a user specified a CPU with LLC_HOST_CPU,
it was used to compile the kernel library and to load it
at link time, but program bitcode compilation still
happened with autodetected CPU. This could cause mismatch
between the library and program bitcodes, resulting in
mysterious errors.

This commit changes CMake variable for autodetected CPU
to LLC_HOST_CPU_AUTO, and leaves LLC_HOST_CPU for the
user. If it's defined and not equal to LLC_HOST_CPU_AUTO,
it's always used for compilation (including program BC).