Add image callbacks to device->ops
There were a few places in pocl where images were assumed
to be always accessible the same way as buffers. This is
not always true with devices like HSA, GPUs and similar.

This separates the entire image handling into their own
driver callbacks, so drivers can deal with images their
own hardware-specific way. Drivers that don't want to
implement images can simply omit the callbacks.

This is mostly an internal API change an this point, since
the only pocl drivers that implement images are CPU drivers.