Add group id as hidden arguments to the WG function
This is instead of the previous way of storing it in the execution
context struct. It produces slightly more efficient and clean code as
there is no need for struct deref to get the WG id.

Also modify example1 so it smoke tests multiple WG execution to catch
errors with WG ids early.