Get rid of cl_command_run->tmp_dir
This was initialized with kernel's cache directory,
and used as a cache key. This commit replaces it
with a per-device array of hashes in kernel metadata.

This has the advantage that it can be calculated once, at
clBuildProgram, and stays valid. Also there's no need to
handle strings.

Note that the kernel hash does NOT include local WG sizes,
which some driver require for cache keys, so this is handled
separately in driver code.