Convert SubBuffers to Buffers in the clEnqueue* API calls
This reduces SubBuffers to ranges inside a Buffer, with
offset and size. There are no allocations (device_ptrs) for
SubBuffers, no host_ptr memory or anything else. The advantage
is that it makes the driver layer simpler (most drivers did not
handle subbuffers correctly anyway), the disadvantage is that
certain optimizations are not possible. This can be fixed later
though, in a more generic manner (not limited to SubBuffers).