Create global offset 0 specialized WG functions.
It seems to be rare to use the global offset in kernel
commands, which means useless address computation which
complicates AA and even incures a little runtime overhead.

This creates a special version of the WG which assumes the
global offset is 0 and uses that in case that's the case.