matrix1: fixes to the test case
Fix the case when max local size is larger than global. Also fix
a div by zero due to an illegal assertion. The div by zero got
triggered if the local wg is larger than matrix size. It just
gets silenced by the FPE handler which is installed in case any
of the CPU devices is built in.