Fix tests for Address Sanitizer builds
Unfortunately the way CMake tests work, if they're given
a pass/fail expression, they don't check for exit status.
This was causing some false negatives with ASan (test was
returning with 1, but CMake reported it as pass because
the pass expression was present in output).

This changes the behaviour to only use pass/fail expressions
when all sanitizers are disabled.