Don't add AlwaysInline to kernels if we're compiling with -cl-disable-opt
FlattenGlobals.cc by default adds the AlwaysInline flag also to
kernels, causing them to be inlined into workgroup launchers. This
makes debugging much harder since GDB sees the source for the kernel
itself but not for the workgroup launchers with the inlined kernel.

This patch alters FlattenGlobals to ignore kernels which have
OptNone flag (-O0 or -cl-disable-opt); this makes the code slightly
slower but debugging much more useful.