AMD SDK 3.0: fix SobelFilter
with C99 and OpenCL, if conversion from float to int
without saturation is called with argument outside integer
range, behaviour is unspecified.

SobelFilter uses convert_uchar4(float4) in OpenCL and simple
typecast in reference C code, both of which resulted in
code truncating floats to uchar range. Until LLVM 11, which
now generates x86 instructions which cause saturation.

It seems it's not possible to specify non-saturating conversion
in OpenCL, so this patch changes both the kernel code and the
reference C code to convert with saturation.